+++ localization-enUS.lua
@@ -7,6 +7,7 @@ L.T["COUNT_FORMAT_PERCENT"] = "Percentage";
 L.T["COUNT_FORMAT_ABSOLUTE"] = "Absolute";
 L.T["COUNT_FORMAT_NONE"] = "None";
 L.T["CREATE_GROUP"] = "Create New Group";
+L.T["CUMULATE_LOWER_QUALITIES"] = "Cumulate Items of Higher Quality";
 L.T["DRAG_ITEM_OR_SET_ITEM_ID"] = "Place an item into the field\nor provide an item ID.";
 L.T["DEFINE_STOCK_GOALS"] = "Set your desired stock goals.";
 L.T["FULLY_STOCKED"] = "Stocked";